django-backend/requirements.txt
Django==4.2.8
psycopg2-binary==2.9.9
python-dotenv==1.0.0
faker==20.1.0
factory-boy==3.3.0
django-extensions==3.2.3
